add doc test
