+++ subxt/src/rpc/types.rs
@@ -233,70 +233,6 @@ into_block_number!(u8 u16 u32 u64);
 /// Arbitrary properties defined in the chain spec as a JSON object.
 pub type SystemProperties = serde_json::Map<String, serde_json::Value>;
 
-/// taken from original type [`pallet_transaction_payment::types::FeeDetails<u128>`]
-#[derive(Encode, Decode, Debug, Clone, Eq, PartialEq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct FeeDetails {
-    /// The minimum fee for a transaction to be included in a block.
-    pub inclusion_fee: Option<InclusionFee>,
-    /// tip
-    pub tip: u128,
-}
-
-/// taken from original type [`pallet_transaction_payment::types::InclusionFee<u128>`]
-/// The base fee and adjusted weight and length fees constitute the _inclusion fee_.
-#[derive(Encode, Decode, Debug, Clone, Eq, PartialEq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct InclusionFee {
-    /// minimum amount a user pays for a transaction.
-    pub base_fee: u128,
-    /// amount paid for the encoded length (in bytes) of the transaction.
-    pub len_fee: u128,
-    ///
-    /// - `targeted_fee_adjustment`: This is a multiplier that can tune the final fee based on the
-    ///   congestion of the network.
-    /// - `weight_fee`: This amount is computed based on the weight of the transaction. Weight
-    /// accounts for the execution time of a transaction.
-    ///
-    /// adjusted_weight_fee = targeted_fee_adjustment * weight_fee
-    pub adjusted_weight_fee: u128,
-}
-
-/// taken from original type [`pallet_transaction_payment::types::RuntimeDispatchInfo<u128,Weight>`]
-#[derive(Debug, Encode, Decode, Eq, PartialEq, Copy, Clone)]
-pub struct RuntimeDispatchInfo {
-    /// Weight of this dispatch.
-    pub weight: Weight,
-    /// Class of this dispatch.
-    pub class: DispatchClass,
-    /// The inclusion fee of this dispatch.
-    ///
-    /// This does not include a tip or anything else
-    pub partial_fee: u128,
-}
-
-/// taken from original type [`sp_weights::weight_v2::Weight`]
-#[derive(Debug, Encode, Decode, Eq, PartialEq, Copy, Clone, Default)]
-pub struct Weight {
-    #[codec(compact)]
-    /// The weight of computational time used based on some reference hardware.
-    pub ref_time: u64,
-    #[codec(compact)]
-    /// The weight of storage space used by proof of validity.
-    pub proof_size: u64,
-}
-
-/// taken from original type [`frame_support::dispatch::DispatchClass`]
-#[derive(Debug, Encode, Decode, Eq, PartialEq, Copy, Clone)]
-pub enum DispatchClass {
-    /// A normal dispatch.
-    Normal,
-    /// An operational dispatch.
-    Operational,
-    /// A mandatory dispatch.
-    Mandatory,
-}
-
 /// Possible transaction status events.
 ///
 /// # Note
