+++ subxt/src/tx/tx_progress.rs
@@ -415,7 +415,7 @@ mod test {
     use futures::Stream;
 
     use crate::{
-        client::{OfflineClientT, OnlineClientT},
+        client::{MockClient, OfflineClientT, OnlineClientT},
         config::{
             extrinsic_params::BaseExtrinsicParams,
             polkadot::{PlainTip, PolkadotConfig},
@@ -429,15 +429,23 @@ mod test {
 
     use serde_json::value::RawValue;
 
+    type MockTxProgress = TxProgress<SubstrateConfig, MockClient>;
+    type MockHash = <WithExtrinsicParams<
+        SubstrateConfig,
+        BaseExtrinsicParams<SubstrateConfig, PlainTip>,
+    > as Config>::Hash;
+    type MockSubstrateTxStatus = SubstrateTxStatus<MockHash, MockHash>;
+
+    /// a mock client to satisfy trait bounds in tests
     #[derive(Clone, Debug)]
     struct MockClient;
 
-    impl OfflineClientT<PolkadotConfig> for MockClient {
+    impl OfflineClientT<SubstrateConfig> for MockClient {
         fn metadata(&self) -> crate::Metadata {
             panic!("just a mock impl to satisfy trait bounds")
         }
 
-        fn genesis_hash(&self) -> <PolkadotConfig as crate::Config>::Hash {
+        fn genesis_hash(&self) -> <SubstrateConfig as crate::Config>::Hash {
             panic!("just a mock impl to satisfy trait bounds")
         }
 
@@ -446,15 +454,8 @@ mod test {
         }
     }
 
-    type MockTxProgress = TxProgress<PolkadotConfig, MockClient>;
-    type MockHash = <WithExtrinsicParams<
-        SubstrateConfig,
-        BaseExtrinsicParams<SubstrateConfig, PlainTip>,
-    > as Config>::Hash;
-    type MockSubstrateTxStatus = SubstrateTxStatus<MockHash, MockHash>;
-
-    impl OnlineClientT<PolkadotConfig> for MockClient {
-        fn rpc(&self) -> &crate::rpc::Rpc<PolkadotConfig> {
+    impl OnlineClientT<SubstrateConfig> for MockClient {
+        fn rpc(&self) -> &crate::rpc::Rpc<SubstrateConfig> {
             panic!("just a mock impl to satisfy trait bounds")
         }
     }
